Update README.md
Kyushimj
This code was originally leaked from an unknown source but most of it was scuffed so i decided to make it now scuffed.

for skids
either download main.py and webserver.py from here or just make the files on replit and paste in the code. on the sidebar next to your files find a lock icon and and add a new secret. for the key name it exactly : DISCORD_BOT_SECRET then in the value field paste in your bot's token. after that press add new secret, go back to your files and you can change the settings to your liking like:

bot prefix
real roblox link (hyperlink)
phis roblox link do not change any of the code unless you know for sure what you're doing.
RUN 24/7 Go to uptimerobot.com, make a new account and add a monitor. name it whatever you want, set it to HTTPs, and paste in the link you get in the little white window that pops up when you run your bot's code. look on youtube for some configurations for the uptimerobot and you're set to go. your bot now runs 24/7 without problems